+++ source/Public/Disable-SqlDscAudit.ps1
@@ -1,22 +1,22 @@
 <#
     .SYNOPSIS
-        Removes a server audit.
+        Disables a server audit.
 
     .PARAMETER ServerObject
         Specifies current server connection object.
 
     .PARAMETER AuditObject
-        Specifies a audit object to remove.
+        Specifies a audit object to disable.
 
     .PARAMETER Name
-        Specifies the name of the server audit to be removed.
+        Specifies the name of the server audit to be disabled.
 
     .PARAMETER Force
-        Specifies that the audit should be removed with out any confirmation.
+        Specifies that the audit should be disabled with out any confirmation.
 
     .PARAMETER Refresh
         Specifies that the **ServerObject**'s audits should be refreshed before
-        trying removing the audit object. This is helpful when audits could have
+        trying to disable the audit object. This is helpful when audits could have
         been modified outside of the **ServerObject**, for example through T-SQL.
         But on instances with a large amount of audits it might be better to make
         sure the **ServerObject** is recent enough, or pass in **AuditObject**.
@@ -24,7 +24,7 @@
     .OUTPUTS
         None.
 #>
-function Remove-SqlDscAudit
+function Disable-SqlDscAudit
 {
     [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseSyntacticallyCorrectExamples', '', Justification = 'Because the rule does not yet support parsing the code when a parameter type is not available. The ScriptAnalyzer rule UseSyntacticallyCorrectExamples will always error in the editor due to https://github.com/indented-automation/Indented.ScriptAnalyzerRules/issues/8.')]
     [OutputType()]
@@ -75,7 +75,7 @@ function Remove-SqlDscAudit
             $PSCmdlet.ThrowTerminatingError(
                 [System.Management.Automation.ErrorRecord]::new(
                     $missingDatabaseMessage,
-                    'RSDA0001', # cspell: disable-line
+                    'DSDA0001', # cspell: disable-line
                     [System.Management.Automation.ErrorCategory]::InvalidOperation,
                     $DatabaseName
                 )
@@ -83,16 +83,12 @@ function Remove-SqlDscAudit
         }
     }
 
-    $verboseDescriptionMessage = $script:localizedData.Audit_Remove_ShouldProcessVerboseDescription -f $AuditObject.Name, $AuditObject.Parent.InstanceName
-    $verboseWarningMessage = $script:localizedData.Audit_Remove_ShouldProcessVerboseWarning -f $AuditObject.Name
-    $captionMessage = $script:localizedData.Audit_Remove_ShouldProcessCaption
+    $verboseDescriptionMessage = $script:localizedData.Audit_Disable_ShouldProcessVerboseDescription -f $AuditObject.Name, $AuditObject.Parent.InstanceName
+    $verboseWarningMessage = $script:localizedData.Audit_Disable_ShouldProcessVerboseWarning -f $AuditObject.Name
+    $captionMessage = $script:localizedData.Audit_Disable_ShouldProcessCaption
 
     if ($PSCmdlet.ShouldProcess($verboseDescriptionMessage, $verboseWarningMessage, $captionMessage))
     {
-        <#
-            If the passed audit object has already been dropped, then we silently
-            do nothing, using the method DropIfExist(), since the job is done.
-        #>
-        $AuditObject.DropIfExist()
+        $AuditObject.Disable()
     }
 }