+++ CHANGELOG.md
@@ -75,6 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `Invoke-SqlDscQuery`
     - `New-SqlDscAudit`
     - `Remove-SqlDscAudit`
+    - `Enable-SqlDscAudit`
+    - `Disable-SqlDscAudit`
   - Support for debugging of integration tests in AppVeyor.
     - Only run for pull requests
   - Add new resource SqlServerAudit.